dom/src/threads/test/test_simpleThread.html

<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Threads (Bug 437152)
-->
<head>
  <title>Test for DOM Worker Threads (Bug 437152)</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=437152">DOM Worker Threads Bug 437152</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

  var worker = new Worker("simpleThread_worker.js");

  worker.addEventListener("message",function(event) {
    is(event.target, worker);
    switch (event.data) {
      case "no-op":
        break;
      case "started":
        is(gotError, true);
        worker.postMessage("no-op");
        worker.postMessage("stop");
        break;
      case "stopped":
        worker.postMessage("no-op");
        SimpleTest.finish();
        break;
      default:
        ok(false, "Unexpected message:" + event.data);
        SimpleTest.finish();
    }
  }, false);

  var gotError = false;
  worker.onerror = function(event) {
    is(event.target, worker);
    is(event.data,
       '[JavaScript Error: "uncaught exception: Bad message: asdf"]');
    gotError = true;

    worker.onerror = function(otherEvent) {
      ok(false, "Worker had an error:" + otherEvent.data);
      SimpleTest.finish();
    }
  };

  worker.postMessage("asdf");
  worker.postMessage("start");

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
